Add Evergreen link to navbar in header content
This commit updates the header content to include a new link directing users to the Evergreen ILS website. The link is added to the navigation bar under the \"navbar-end\" section, providing quick and easy access to the Evergreen site alongside existing links.

Release-Note: Add Evergreen website link to the navbar for easy access

Signed-off-by: IanSkelskey <ianskelskey@gmail.com>